chore: create cloud-providers domain
