+++ libs/domains/cloud-providers/data-access/README.md
@@ -1,0 +1,7 @@
+# domains-cloud-providers-data-access
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test domains-cloud-providers-data-access` to execute the unit tests via [Jest](https://jestjs.io).
