+++ libs/domains/cloud-providers/feature/README.md
@@ -1,0 +1,7 @@
+# domains-cloud-providers-feature
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test domains-cloud-providers-feature` to execute the unit tests via [Jest](https://jestjs.io).
