+++ libs/domains/cloud-providers/feature/src/lib/hooks/use-cloud-providers/use-cloud-providers.ts
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query'
+import { queries } from '@qovery/state/util-queries'
+
+export function useCloudProviders() {
+  return useQuery({
+    ...queries.cloudProviders.list,
+  })
+}
+
+export default useCloudProviders
