+++ libs/domains/clusters/feature/src/index.ts
@@ -1,6 +1,5 @@
 export * from './lib/cluster-type/cluster-type'
 export * from './lib/cluster-delete-modal/cluster-delete-modal'
-export * from './lib/hooks/use-cloud-provider-features/use-cloud-provider-features'
 export * from './lib/hooks/use-cluster-cloud-provider-info/use-cluster-cloud-provider-info'
 export * from './lib/hooks/use-cluster-routing-table/use-cluster-routing-table'
 export * from './lib/hooks/use-cluster-logs/use-cluster-logs'
