+++ libs/domains/organizations/feature/src/index.ts
@@ -12,11 +12,9 @@ export * from './lib/hooks/use-available-helm-repositories/use-available-helm-re
 export * from './lib/hooks/use-available-roles/use-available-roles'
 export * from './lib/hooks/use-billing-info/use-billing-info'
 export * from './lib/hooks/use-branches/use-branches'
-export * from './lib/hooks/use-cloud-provider-credentials/use-cloud-provider-credentials'
 export * from './lib/hooks/use-container-registries/use-container-registries'
 export * from './lib/hooks/use-container-registry/use-container-registry'
 export * from './lib/hooks/use-create-api-token/use-create-api-token'
-export * from './lib/hooks/use-create-cloud-provider-credential/use-create-cloud-provider-credential'
 export * from './lib/hooks/use-create-container-registry/use-create-container-registry'
 export * from './lib/hooks/use-create-custom-role/use-create-custom-role'
 export * from './lib/hooks/use-create-helm-repository/use-create-helm-repository'
@@ -27,7 +25,6 @@ export * from './lib/hooks/use-credit-cards/use-credit-cards'
 export * from './lib/hooks/use-current-cost/use-current-cost'
 export * from './lib/hooks/use-custom-role/use-custom-role'
 export * from './lib/hooks/use-delete-api-token/use-delete-api-token'
-export * from './lib/hooks/use-delete-cloud-provider-credential/use-delete-cloud-provider-credential'
 export * from './lib/hooks/use-delete-container-registry/use-delete-container-registry'
 export * from './lib/hooks/use-delete-credit-card/use-delete-credit-card'
 export * from './lib/hooks/use-delete-custom-role/use-delete-custom-role'
@@ -37,7 +34,6 @@ export * from './lib/hooks/use-delete-member/use-delete-member'
 export * from './lib/hooks/use-delete-organization/use-delete-organization'
 export * from './lib/hooks/use-delete-webhook/use-delete-webhook'
 export * from './lib/hooks/use-edit-billing-info/use-edit-billing-info'
-export * from './lib/hooks/use-edit-cloud-provider-credential/use-edit-cloud-provider-credential'
 export * from './lib/hooks/use-edit-container-registry/use-edit-container-registry'
 export * from './lib/hooks/use-edit-custom-role/use-edit-custom-role'
 export * from './lib/hooks/use-edit-helm-repository/use-edit-helm-repository'
