+++ libs/pages/clusters/src/lib/feature/page-clusters-create-feature/step-general-feature/step-general-feature.tsx
@@ -2,8 +2,8 @@ import { useEffect } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 import { useDispatch, useSelector } from 'react-redux'
 import { useNavigate, useParams } from 'react-router-dom'
+import { useCloudProviderCredentials } from '@qovery/domains/cloud-providers/feature'
 import { fetchCloudProvider, getClusterState } from '@qovery/domains/organization'
-import { useCloudProviderCredentials } from '@qovery/domains/organizations/feature'
 import { type ClusterGeneralData } from '@qovery/shared/interfaces'
 import { CLUSTERS_CREATION_RESOURCES_URL, CLUSTERS_CREATION_URL, CLUSTERS_URL } from '@qovery/shared/routes'
 import { FunnelFlowBody, FunnelFlowHelpCard } from '@qovery/shared/ui'
