+++ tsconfig.base.json
@@ -16,6 +16,8 @@
     "baseUrl": ".",
     "paths": {
       "@qovery/domains/application": ["libs/domains/application/src/index.ts"],
+      "@qovery/domains/cloud-providers/data-access": ["libs/domains/cloud-providers/data-access/src/index.ts"],
+      "@qovery/domains/cloud-providers/feature": ["libs/domains/cloud-providers/feature/src/index.ts"],
       "@qovery/domains/clusters/data-access": ["libs/domains/clusters/data-access/src/index.ts"],
       "@qovery/domains/clusters/feature": ["libs/domains/clusters/feature/src/index.ts"],
       "@qovery/domains/database": ["libs/domains/database/src/index.ts"],
